refactor(frontend): rebuild the Postgres/NoSql/Redis inspector modals on the shared components
Deduplicates the three near-identical modals (~3700 lines down to ~800 plus
shared components): identical shell/tabs/styles/fetch-retry/cheat-sheet/query
console now live in features/nodes/components. The genuinely per-DB parts
(simulation topologies, query handlers, explorer shapes) stay per node.

Behavior changes (approved): the NoSql explorer gains the Retry button the
other two already had, and the Postgres/NoSql startup auto-retry timers no
longer leak or stack (they adopt the Redis retryTimerRef pattern via
useExplorerData).
